Updated for Visual Studio

+++ gaze_data/src/Ray.cpp
@@ -14,8 +14,8 @@ CVector<2> CRay::PointOfShortestDistance( const CRay &other ) const
 void CRay::Render( CImage &img, const cv::Scalar &colorPoint, const cv::Scalar &colorLine, double dLength, int iRadius, int iPointThickness, int iLineThickness ) const
 {
 	CVector<3> vec3Endp = ( *this )( dLength );
-	cv::Point pt1( m_vec3Origin[ 0 ] * img.matImage.cols, m_vec3Origin[ 1 ] * img.matImage.rows );
-	cv::Point pt2( vec3Endp[ 0 ] * img.matImage.cols, vec3Endp[ 1 ] * img.matImage.rows );
+	cv::Point pt1( (int) ( m_vec3Origin[ 0 ] * img.matImage.cols ), (int) ( m_vec3Origin[ 1 ] * img.matImage.rows ) );
+	cv::Point pt2( (int) ( vec3Endp[ 0 ] * img.matImage.cols ), (int) ( vec3Endp[ 1 ] * img.matImage.rows ) );
 	cv::circle( img.matImage, pt1, iRadius, colorPoint, iPointThickness );
 	cv::line( img.matImage, pt1, pt2, colorLine, iLineThickness );
 }
